+++ supertokens_python/recipe/thirdpartypasswordless/recipeimplementation/implementation.py
@@ -15,11 +15,20 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Dict, List, Union
 
-from ...passwordless.interfaces import (CreateCodeResult,
-                                        CreateNewCodeForDeviceResult,
-                                        DeviceType, RevokeAllCodesResult,
-                                        RevokeCodeResult, UpdateUserResult)
-from ...thirdparty.interfaces import SignInUpResult
+from ...passwordless.interfaces import (
+    ConsumeCodeExpiredUserInputCodeError,
+    ConsumeCodeIncorrectUserInputCodeError,
+    ConsumeCodeRestartFlowError, CreateCodeOkResult,
+    CreateNewCodeForDeviceOkResult,
+    CreateNewCodeForDeviceRestartFlowError,
+    CreateNewCodeForDeviceUserInputCodeAlreadyUsedError,
+    DeleteUserInfoOkResult, DeleteUserInfoUnknownUserIdError,
+    DeviceType, RevokeAllCodesOkResult, RevokeCodeOkResult,
+    UpdateUserEmailAlreadyExistsError, UpdateUserOkResult,
+    UpdateUserPhoneNumberAlreadyExistsError,
+    UpdateUserUnknownUserIdError)
+from ...passwordless.interfaces import ConsumeCodeOkResult as PasswordlessConsumeCodeOkResult
+from ...thirdparty.interfaces import SignInUpFieldError, SignInUpOkResult
 
 if TYPE_CHECKING:
     from supertokens_python.querier import Querier
@@ -29,11 +38,7 @@ from supertokens_python.recipe.passwordless.recipe_implementation import \
 from supertokens_python.recipe.thirdparty.recipe_implementation import \
     RecipeImplementation as ThirdPartyImplementation
 
-from ..interfaces import (ConsumeCodeExpiredUserInputCodeErrorResult,
-                          ConsumeCodeIncorrectUserInputCodeErrorResult,
-                          ConsumeCodeOkResult,
-                          ConsumeCodeRestartFlowErrorResult, ConsumeCodeResult,
-                          RecipeInterface)
+from ..interfaces import RecipeInterface, ConsumeCodeOkResult
 from ..types import User
 from .passwordless_recipe_implementation import \
     RecipeImplementation as DerivedPasswordlessImplementation
@@ -61,6 +66,8 @@ class RecipeImplementation(RecipeInterface):
         self.pless_revoke_all_codes = passwordless_implementation.revoke_all_codes
         self.pless_revoke_code = passwordless_implementation.revoke_code
         self.pless_update_user = passwordless_implementation.update_user
+        self.pless_delete_email_for_user = passwordless_implementation.delete_email_for_user
+        self.pless_delete_phone_number_for_user = passwordless_implementation.delete_phone_number_for_user
 
         derived_pless = DerivedPasswordlessImplementation(self)
         passwordless_implementation.get_user_by_id = derived_pless.get_user_by_id
@@ -76,6 +83,8 @@ class RecipeImplementation(RecipeInterface):
         passwordless_implementation.revoke_all_codes = derived_pless.revoke_all_codes
         passwordless_implementation.revoke_code = derived_pless.revoke_code
         passwordless_implementation.update_user = derived_pless.update_user
+        passwordless_implementation.delete_email_for_user = derived_pless.delete_email_for_user
+        passwordless_implementation.delete_phone_number_for_user = derived_pless.delete_phone_number_for_user
 
         self.tp_get_user_by_id = None
         self.tp_get_users_by_email = None
@@ -137,7 +146,7 @@ class RecipeImplementation(RecipeInterface):
         return User(user_id=tp_user.user_id, email=tp_user.email, time_joined=tp_user.time_joined, third_party_info=tp_user.third_party_info, phone_number=None)
 
     async def thirdparty_sign_in_up(self, third_party_id: str, third_party_user_id: str, email: str,
-                                    email_verified: bool, user_context: Dict[str, Any]) -> SignInUpResult:
+                                    email_verified: bool, user_context: Dict[str, Any]) -> Union[SignInUpOkResult, SignInUpFieldError]:
         if self.tp_sign_in_up is None:
             raise Exception("No thirdparty provider configured")
         return await self.tp_sign_in_up(third_party_id, third_party_user_id, email, email_verified, user_context)
@@ -153,13 +162,13 @@ class RecipeImplementation(RecipeInterface):
                           email: Union[None, str],
                           phone_number: Union[None, str],
                           user_input_code: Union[None, str],
-                          user_context: Dict[str, Any]) -> CreateCodeResult:
+                          user_context: Dict[str, Any]) -> CreateCodeOkResult:
         return await self.pless_create_code(email, phone_number, user_input_code, user_context)
 
     async def create_new_code_for_device(self,
                                          device_id: str,
                                          user_input_code: Union[str, None],
-                                         user_context: Dict[str, Any]) -> CreateNewCodeForDeviceResult:
+                                         user_context: Dict[str, Any]) -> Union[CreateNewCodeForDeviceOkResult, CreateNewCodeForDeviceRestartFlowError, CreateNewCodeForDeviceUserInputCodeAlreadyUsedError]:
         return await self.pless_create_new_code_for_device(device_id, user_input_code, user_context)
 
     async def consume_code(self,
@@ -167,34 +176,48 @@ class RecipeImplementation(RecipeInterface):
                            user_input_code: Union[str, None],
                            device_id: Union[str, None],
                            link_code: Union[str, None],
-                           user_context: Dict[str, Any]) -> ConsumeCodeResult:
+                           user_context: Dict[str, Any]) -> Union[ConsumeCodeOkResult, ConsumeCodeIncorrectUserInputCodeError, ConsumeCodeExpiredUserInputCodeError, ConsumeCodeRestartFlowError]:
         result = await self.pless_consume_code(pre_auth_session_id, user_input_code, device_id, link_code, user_context)
-
-        if result.is_ok:
-            if result.user is None or result.created_new_user is None:
-                raise Exception("Should never come here")
-            return ConsumeCodeOkResult(result.created_new_user, User(result.user.user_id, result.user.email, result.user.phone_number, None, result.user.time_joined))
-        if result.is_expired_user_input_code_error:
-            if result.failed_code_input_attempt_count is None or result.maximum_code_input_attempts is None:
-                raise Exception("Should never come here")
-            return ConsumeCodeExpiredUserInputCodeErrorResult(result.failed_code_input_attempt_count, result.maximum_code_input_attempts)
-        if result.is_incorrect_user_input_code_error:
-            if result.failed_code_input_attempt_count is None or result.maximum_code_input_attempts is None:
-                raise Exception("Should never come here")
-            return ConsumeCodeIncorrectUserInputCodeErrorResult(result.failed_code_input_attempt_count, result.maximum_code_input_attempts)
-
-        # restart flow error
-        return ConsumeCodeRestartFlowErrorResult()
+        if isinstance(result, PasswordlessConsumeCodeOkResult):
+            return ConsumeCodeOkResult(
+                result.created_new_user,
+                User(result.user.user_id, result.user.email, result.user.phone_number, None, result.user.time_joined)
+            )
+        return result
 
     async def update_passwordless_user(self, user_id: str,
-                                       email: Union[str, None], phone_number: Union[str, None], user_context: Dict[str, Any]) -> UpdateUserResult:
+                                       email: Union[str, None], phone_number: Union[str, None], user_context: Dict[str, Any]) -> Union[UpdateUserOkResult, UpdateUserUnknownUserIdError, UpdateUserEmailAlreadyExistsError, UpdateUserPhoneNumberAlreadyExistsError]:
+        user = await self.get_user_by_id(user_id, user_context)
+        if user is None:
+            return UpdateUserUnknownUserIdError()
+        if user.third_party_info is not None:
+            raise Exception(
+                "Cannot update passwordless user info of a user who signed up using third party login.")
         return await self.pless_update_user(user_id, email, phone_number, user_context)
 
+    async def delete_email_for_passwordless_user(self, user_id: str, user_context: Dict[str, Any]) -> Union[DeleteUserInfoOkResult, DeleteUserInfoUnknownUserIdError]:
+        user = await self.get_user_by_id(user_id, user_context)
+        if user is None:
+            return DeleteUserInfoUnknownUserIdError()
+        if user.third_party_info is not None:
+            raise Exception(
+                "Cannot update passwordless user info of a user who signed up using third party login.")
+        return await self.pless_delete_email_for_user(user_id, user_context)
+
+    async def delete_phone_number_for_user(self, user_id: str, user_context: Dict[str, Any]) -> Union[DeleteUserInfoOkResult, DeleteUserInfoUnknownUserIdError]:
+        user = await self.get_user_by_id(user_id, user_context)
+        if user is None:
+            return DeleteUserInfoUnknownUserIdError()
+        if user.third_party_info is not None:
+            raise Exception(
+                "Cannot update passwordless user info of a user who signed up using third party login.")
+        return await self.pless_delete_phone_number_for_user(user_id, user_context)
+
     async def revoke_all_codes(self,
-                               email: Union[str, None], phone_number: Union[str, None], user_context: Dict[str, Any]) -> RevokeAllCodesResult:
+                               email: Union[str, None], phone_number: Union[str, None], user_context: Dict[str, Any]) -> RevokeAllCodesOkResult:
         return await self.pless_revoke_all_codes(email, phone_number, user_context)
 
-    async def revoke_code(self, code_id: str, user_context: Dict[str, Any]) -> RevokeCodeResult:
+    async def revoke_code(self, code_id: str, user_context: Dict[str, Any]) -> RevokeCodeOkResult:
         return await self.pless_revoke_code(code_id, user_context)
 
     async def list_codes_by_email(self, email: str, user_context: Dict[str, Any]) -> List[DeviceType]: