+++ supertokens_python/utils.py
@@ -15,6 +15,7 @@
 from __future__ import annotations
 
 import asyncio
+import warnings
 from base64 import b64decode, b64encode
 from re import fullmatch
 from time import time
@@ -153,13 +154,23 @@ def find_first_occurrence_in_list(
     return None
 
 
-def execute_in_background(mode: str, func: Callable[[], Coroutine[Any, Any, None]]):
-    if mode == 'wsgi':
+def execute_async(mode: str, func: Callable[[], Coroutine[Any, Any, None]]):
+    real_mode = None
+    try:
+        asyncio.get_running_loop()
+        real_mode = 'asgi'
+    except RuntimeError:
+        real_mode = 'wsgi'
+
+    if mode != real_mode:
+        warnings.warn('Inconsistent mode detected, check if you are using the right asgi / wsgi mode', category=RuntimeWarning)
+
+    if real_mode == 'wsgi':
+        asyncio.run(func())
+    else:
         check_event_loop()
         loop = asyncio.get_event_loop()
-        loop.run_until_complete(func())
-    else:
-        asyncio.create_task(func())
+        loop.create_task(func())
 
 
 def frontend_has_interceptor(request: BaseRequest) -> bool: